Move Packet and PacketHandle classes to Parser namespace
