+++ src/Microsoft.Data.SqlClient/tests/FunctionalTests/Microsoft.Data.SqlClient.FunctionalTests.csproj
@@ -20,7 +20,7 @@
     <!-- compile them as part of this project. This is not ideal in the slightest, and they    -->
     <!-- should ideally be UnitTests if they need to access internals. But detangling these    -->
     <!-- tests is far beyond the scope of the common project, so it is left as a TODO for now. -->
-    <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Packet.cs" />
+    <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/Packet.cs" />
     <Compile Include="$(RepoRoot)src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/TdsParserStateObject.Multiplexer.cs" />
   </ItemGroup>
